bons10 - Setting up Kafka and kafka topic
Unfortunately, I won't have the time to figure out how to "make it easy to setup". This is a problem that I will not be able to address at this moment.

For now, it's a manual setup of installing, unzipping the tar.gz file and then praying to God that there isn't any PORT issues when running :)